Arrays

+++ OOPTest/src/com/raj/IssueWithDouble.java
@@ -6,8 +6,8 @@ package com.raj;
 public class IssueWithDouble {
 
     public static void main(String[] args) {
-        double d1 = 5.23;
-        double d2 = 1.11;
+        float d1 = (float)5.23;
+        float d2 = (float)1.11;
 
         System.out.println(d1+d2);
     }
